feat: mada a draft of 2nd step
- additionally made creation of event only available as modal
