--- src/app/dashboard/event/create/layout.tsx
@@ -1,7 +1,0 @@
-export default function CreateEventFormLayout({
-  children,
-}: {
-  children: Readonly<React.ReactNode>;
-}) {
-  return <div className="p-6">{children}</div>;
-}
